chore(plugins) cleanup plugins & operator loading
