+++ app/packages/app/src/components/Nav.tsx
@@ -13,8 +13,10 @@ import {
 } from "@fiftyone/components";
 import { ViewBar } from "@fiftyone/core";
 import { OperatorPlacements, types } from "@fiftyone/operators";
+import { PANEL_AREA, PanelArea } from "@fiftyone/spaces";
 import * as fos from "@fiftyone/state";
 import { useRefresh } from "@fiftyone/state";
+import { constants } from "@fiftyone/utilities";
 import { DarkMode, LightMode } from "@mui/icons-material";
 import { Box, useColorScheme } from "@mui/material";
 import React, { Suspense, useMemo } from "react";
@@ -27,6 +29,8 @@ import DatasetSelector from "./DatasetSelector";
 import Teams from "./Teams";
 import type { NavDatasets$key } from "./__generated__/NavDatasets.graphql";
 import type { NavFragment$key } from "./__generated__/NavFragment.graphql";
+
+const { IS_APP_MODE_FIFTYONE } = constants;
 
 const getUseSearch = (fragment: NavDatasets$key) => {
   return (search: string) => {
@@ -130,7 +134,22 @@ const Nav: React.FC<
           </Box>
         </div>
       </Header>
-      {children}
+      <Box
+        sx={{
+          display: "flex",
+          flex: 1,
+          minHeight: 0,
+          overflow: "hidden",
+        }}
+      >
+        {children}
+        {IS_APP_MODE_FIFTYONE && (
+          <PanelArea
+            id={PANEL_AREA.SIDEBAR_RIGHT}
+            resize={{ direction: "left" }}
+          />
+        )}
+      </Box>
       <Analytics fragment={data} />
     </>
   );
