+++ app/packages/core/src/plugins/index.ts
@@ -1,4 +1,3 @@
-import "./OperatorIO";
 import "./SchemaIO";
 import "./histograms";
 import "./modal-sample";
