+++ app/packages/operators/src/OperatorIOComponent/examples/input.json
