+++ app/packages/operators/src/OperatorIOComponent/examples/output.json
