+++ app/packages/operators/src/OperatorIOComponent/examples/plotly.json
