+++ app/packages/operators/src/OperatorIOComponent/examples/smart-form.json
