+++ app/packages/operators/src/index.ts
@@ -37,3 +37,4 @@ export {
   useTriggerPanelEvent,
 } from "./usePanelEvent";
 export { validate } from "./validation";
+export { operatorToIOSchema } from "./OperatorIOComponent/utils";
