+++ app/packages/state/src/hooks/plugins/useSpacesContext.ts
@@ -1,6 +1,8 @@
 import { useRecoilValue } from "recoil";
-import { fieldSchema, State, dataset as datasetAtom } from "../../recoil";
 import { useMemo } from "react";
+import { dataset as datasetAtom } from "../../recoil/dataset";
+import { fieldSchema } from "../../recoil/schema";
+import { State } from "../../recoil/types";
 
 export default function useSpacesContext() {
   const schema = useRecoilValue(fieldSchema({ space: State.SPACE.SAMPLE }));
